Working with fixing problems and adding functionality refactoring Heroku sepeate folders
