+++ .env.example
@@ -1,2 +1,2 @@
 HEROKU_API_KEY=YOUR_API_KEY
-GITHUB_API_KEY=YOUR_API_KEY
+GITHUB_ACCESS_TOKEN=YOUR_API_KEY
