--- api_responses/heroku_invoice_team.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0"?>
-<x1:invoice xmlns:x1="http://example.org/invoices">
- <x1:InvoiceItem xmlns:x1="http://example.org/invoices">
-  <x1:charges_total xmlns:x1="http://example.org/invoices">0</x1:charges_total>
-  <x1:created_at xmlns:x1="http://example.org/invoices">2022-03-01T04:09:09Z</x1:created_at>
-  <x1:credits_total xmlns:x1="http://example.org/invoices"/>
-  <x1:id xmlns:x1="http://example.org/invoices">9aa53ffa-ce4f-487d-ba74-2fffb3c98e04</x1:id>
-  <x1:number xmlns:x1="http://example.org/invoices">64229959</x1:number>
-  <x1:period_start xmlns:x1="http://example.org/invoices">2022-03-01</x1:period_start>
-  <x1:period_end xmlns:x1="http://example.org/invoices">2022-03-16</x1:period_end>
-  <x1:state xmlns:x1="http://example.org/invoices">0</x1:state>
-  <x1:total xmlns:x1="http://example.org/invoices">0</x1:total>
-  <x1:updated_at xmlns:x1="http://example.org/invoices">2022-03-16T04:14:36Z</x1:updated_at>
-  <x1:addons_total xmlns:x1="http://example.org/invoices">0</x1:addons_total>
-  <x1:database_total xmlns:x1="http://example.org/invoices">0</x1:database_total>
-  <x1:dyno_units xmlns:x1="http://example.org/invoices">0</x1:dyno_units>
-  <x1:platform_total xmlns:x1="http://example.org/invoices">0</x1:platform_total>
-  <x1:payment_status xmlns:x1="http://example.org/invoices">Paid</x1:payment_status>
-  <x1:weighted_dyno_hours xmlns:x1="http://example.org/invoices">0</x1:weighted_dyno_hours>
- </x1:InvoiceItem>
- <x1:InvoiceItem xmlns:x1="http://example.org/invoices">
-  <x1:charges_total xmlns:x1="http://example.org/invoices">0</x1:charges_total>
-  <x1:created_at xmlns:x1="http://example.org/invoices">2022-02-22T04:17:57Z</x1:created_at>
-  <x1:credits_total xmlns:x1="http://example.org/invoices"/>
-  <x1:id xmlns:x1="http://example.org/invoices">c1f7f8b3-b063-4001-aebd-7a7600ebf206</x1:id>
-  <x1:number xmlns:x1="http://example.org/invoices">62986019</x1:number>
-  <x1:period_start xmlns:x1="http://example.org/invoices">2022-02-01</x1:period_start>
-  <x1:period_end xmlns:x1="http://example.org/invoices">2022-03-01</x1:period_end>
-  <x1:state xmlns:x1="http://example.org/invoices">1</x1:state>
-  <x1:total xmlns:x1="http://example.org/invoices">0</x1:total>
-  <x1:updated_at xmlns:x1="http://example.org/invoices">2022-03-01T14:22:08Z</x1:updated_at>
-  <x1:addons_total xmlns:x1="http://example.org/invoices">0</x1:addons_total>
-  <x1:database_total xmlns:x1="http://example.org/invoices">0</x1:database_total>
-  <x1:dyno_units xmlns:x1="http://example.org/invoices">0</x1:dyno_units>
-  <x1:platform_total xmlns:x1="http://example.org/invoices">0</x1:platform_total>
-  <x1:payment_status xmlns:x1="http://example.org/invoices">Paid</x1:payment_status>
-  <x1:weighted_dyno_hours xmlns:x1="http://example.org/invoices">0</x1:weighted_dyno_hours>
- </x1:InvoiceItem>
-</x1:invoice>
