+++ api_responses_json/google_skus.json
@@ -1,0 +1,135 @@
+[
+    {
+        "sku_name": "services\/0069-3716-5463\/skus\/00C0-AE90-D5AE",
+        "sku_id": "00C0-AE90-D5AE",
+        "sku_description": "Custom Commit Plan QUser",
+        "sku_provider_name": "Qubole",
+        "sku_service_name": "Qubole Data Service",
+        "sku_resource": "ApplicationServices",
+        "sku_group": "Qubole",
+        "sku_usage_type": "OnDemand",
+        "sku_effective_time": "18:42:26",
+        "sku_usage_unit": "mo",
+        "sku_usage_unit_description": "month",
+        "sku_base_unit": "s",
+        "sku_base_unit_description": "second",
+        "sku_base_unit_conversion_factor": 2674800,
+        "sku_display_quantity": 1,
+        "sku_start_usage_amount": 0,
+        "sku_unit_price": "USD"
+    },
+    {
+        "sku_name": "services\/0069-3716-5463\/skus\/1992-9BAC-A2C0",
+        "sku_id": "1992-9BAC-A2C0",
+        "sku_description": "Custom Commit Plan QCU",
+        "sku_provider_name": "Qubole",
+        "sku_service_name": "Qubole Data Service",
+        "sku_resource": "ApplicationServices",
+        "sku_group": "Qubole",
+        "sku_usage_type": "OnDemand",
+        "sku_effective_time": "18:42:26",
+        "sku_usage_unit": "h",
+        "sku_usage_unit_description": "hour",
+        "sku_base_unit": "s",
+        "sku_base_unit_description": "second",
+        "sku_base_unit_conversion_factor": 3600,
+        "sku_display_quantity": 1,
+        "sku_start_usage_amount": 0,
+        "sku_unit_price": "USD"
+    },
+    {
+        "sku_name": "services\/0069-3716-5463\/skus\/430A-4B21-3852",
+        "sku_id": "430A-4B21-3852",
+        "sku_description": "Essential Plan QCU",
+        "sku_provider_name": "Qubole",
+        "sku_service_name": "Qubole Data Service",
+        "sku_resource": "ApplicationServices",
+        "sku_group": "Qubole",
+        "sku_usage_type": "OnDemand",
+        "sku_effective_time": "18:42:26",
+        "sku_usage_unit": "h",
+        "sku_usage_unit_description": "hour",
+        "sku_base_unit": "s",
+        "sku_base_unit_description": "second",
+        "sku_base_unit_conversion_factor": 3600,
+        "sku_display_quantity": 1,
+        "sku_start_usage_amount": 0,
+        "sku_unit_price": "USD"
+    },
+    {
+        "sku_name": "services\/0069-3716-5463\/skus\/71ED-0CFF-B099",
+        "sku_id": "71ED-0CFF-B099",
+        "sku_description": "Custom Commit Plan QUser - Tiered Option 3",
+        "sku_provider_name": "Qubole",
+        "sku_service_name": "Qubole Data Service",
+        "sku_resource": "ApplicationServices",
+        "sku_group": "Qubole",
+        "sku_usage_type": "OnDemand",
+        "sku_effective_time": "18:42:26",
+        "sku_usage_unit": "mo",
+        "sku_usage_unit_description": "month",
+        "sku_base_unit": "s",
+        "sku_base_unit_description": "second",
+        "sku_base_unit_conversion_factor": 2674800,
+        "sku_display_quantity": 1,
+        "sku_start_usage_amount": 0,
+        "sku_unit_price": "USD"
+    },
+    {
+        "sku_name": "services\/0069-3716-5463\/skus\/8127-822E-C8F6",
+        "sku_id": "8127-822E-C8F6",
+        "sku_description": "Essential Plan QUser",
+        "sku_provider_name": "Qubole",
+        "sku_service_name": "Qubole Data Service",
+        "sku_resource": "ApplicationServices",
+        "sku_group": "Qubole",
+        "sku_usage_type": "OnDemand",
+        "sku_effective_time": "18:42:26",
+        "sku_usage_unit": "mo",
+        "sku_usage_unit_description": "month",
+        "sku_base_unit": "s",
+        "sku_base_unit_description": "second",
+        "sku_base_unit_conversion_factor": 2674800,
+        "sku_display_quantity": 1,
+        "sku_start_usage_amount": 0,
+        "sku_unit_price": "USD"
+    },
+    {
+        "sku_name": "services\/0069-3716-5463\/skus\/C41F-D592-7A2A",
+        "sku_id": "C41F-D592-7A2A",
+        "sku_description": "Custom Commit Plan QUser - Tiered Option 1",
+        "sku_provider_name": "Qubole",
+        "sku_service_name": "Qubole Data Service",
+        "sku_resource": "ApplicationServices",
+        "sku_group": "Qubole",
+        "sku_usage_type": "OnDemand",
+        "sku_effective_time": "18:42:26",
+        "sku_usage_unit": "mo",
+        "sku_usage_unit_description": "month",
+        "sku_base_unit": "s",
+        "sku_base_unit_description": "second",
+        "sku_base_unit_conversion_factor": 2674800,
+        "sku_display_quantity": 1,
+        "sku_start_usage_amount": 0,
+        "sku_unit_price": "USD"
+    },
+    {
+        "sku_name": "services\/0069-3716-5463\/skus\/EC90-9230-B2C6",
+        "sku_id": "EC90-9230-B2C6",
+        "sku_description": "Custom Commit Plan QUser - Tiered Option 2",
+        "sku_provider_name": "Qubole",
+        "sku_service_name": "Qubole Data Service",
+        "sku_resource": "ApplicationServices",
+        "sku_group": "Qubole",
+        "sku_usage_type": "OnDemand",
+        "sku_effective_time": "18:42:26",
+        "sku_usage_unit": "mo",
+        "sku_usage_unit_description": "month",
+        "sku_base_unit": "s",
+        "sku_base_unit_description": "second",
+        "sku_base_unit_conversion_factor": 2674800,
+        "sku_display_quantity": 1,
+        "sku_start_usage_amount": 0,
+        "sku_unit_price": "USD"
+    }
+]
