+++ api_responses_json/heroku_invoice.json
@@ -1,0 +1,26 @@
+[
+    {
+        "charges_total": 0,
+        "created_at": "2022-02-17T02:45:30Z",
+        "credits_total": 0,
+        "id": "22e1c650-291e-48b7-b462-4bbd9eba4a0e",
+        "number": 62983158,
+        "period_end": "2022-03-01",
+        "period_start": "2022-02-01",
+        "state": 1,
+        "total": 0,
+        "updated_at": "2022-03-01T12:59:11Z"
+    },
+    {
+        "charges_total": 0,
+        "created_at": "2022-03-01T02:44:04Z",
+        "credits_total": 0,
+        "id": "e4bbdb15-40db-4de8-b89f-d2f921766302",
+        "number": 63410419,
+        "period_end": "2022-03-17",
+        "period_start": "2022-03-01",
+        "state": 0,
+        "total": 0,
+        "updated_at": "2022-03-17T02:40:34Z"
+    }
+]
