+++ api_responses_json/heroku_team_invoices.json
@@ -1,0 +1,38 @@
+[
+    {
+        "charges_total": 0,
+        "created_at": "2022-03-01T04:09:09Z",
+        "credits_total": 0,
+        "id": "9aa53ffa-ce4f-487d-ba74-2fffb3c98e04",
+        "number": 64229959,
+        "period_end": "2022-03-17",
+        "period_start": "2022-03-01",
+        "state": 0,
+        "total": 0,
+        "updated_at": "2022-03-17T03:58:33Z",
+        "addons_total": 0,
+        "database_total": 0,
+        "dyno_units": 0,
+        "platform_total": 0,
+        "payment_status": "Paid",
+        "weighted_dyno_hours": 0
+    },
+    {
+        "charges_total": 0,
+        "created_at": "2022-02-22T04:17:57Z",
+        "credits_total": 0,
+        "id": "c1f7f8b3-b063-4001-aebd-7a7600ebf206",
+        "number": 62986019,
+        "period_end": "2022-03-01",
+        "period_start": "2022-02-01",
+        "state": 1,
+        "total": 0,
+        "updated_at": "2022-03-01T14:22:08Z",
+        "addons_total": 0,
+        "database_total": 0,
+        "dyno_units": 0,
+        "platform_total": 0,
+        "payment_status": "Paid",
+        "weighted_dyno_hours": 0
+    }
+]
