+++ api_responses_xml/github-billing-storage-user.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <x1:billing xmlns:x1="http://example.org/billings">
- <x1:days_left_in_billing_cycle xmlns:x1="http://example.org/billings">27</x1:days_left_in_billing_cycle>
+ <x1:days_left_in_billing_cycle xmlns:x1="http://example.org/billings">15</x1:days_left_in_billing_cycle>
  <x1:estimated_paid_storage_for_month xmlns:x1="http://example.org/billings">0</x1:estimated_paid_storage_for_month>
  <x1:estimated_storage_for_month xmlns:x1="http://example.org/billings">0</x1:estimated_storage_for_month>
 </x1:billing>
